requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
fastapi
streamlit
langsmith
langserve
uvicorn